Move KubernetesPortMode enum to sky.utils.kubernetes_enums
